+++ metatensor-core/CHANGELOG.md
@@ -29,6 +29,13 @@ a changelog](https://keepachangelog.com/en/1.1.0/) format. This project follows
 
 - `Labels::set_difference` for finding the set difference between two `Labels`
 
+### metatensor-core Python
+
+#### Added
+
+- `Labels.difference` and `Labels.difference_and_mapping` for finding the set
+  difference between two `Labels`
+
 ## [Version 0.1.12](https://github.com/metatensor/metatensor/releases/tag/metatensor-core-v0.1.12) - 2025-02-17
 
 ### Changed